+++ src/FunFair.BuildCheck.ProjectChecks/Settings/MustNotDefineCodeAnalysisRuleSet.cs
@@ -1,0 +1,9 @@
+using Microsoft.Extensions.Logging;
+
+namespace FunFair.BuildCheck.ProjectChecks.Settings;
+
+public sealed class MustNotDefineCodeAnalysisRuleSet : MustNotDefinePropertyProjectCheckBase
+{
+    public MustNotDefineCodeAnalysisRuleSet(ILogger<MustNotDefineCodeAnalysisRuleSet> logger)
+        : base(propertyName: "CodeAnalysisRuleSet", logger: logger) { }
+}